Fixed handling of unit variants of aliased enums in type NS.
